+++ tests/Feature/Framework/ListeningEvents/Lifecycle/Module/Factory.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Feature\Framework\ListeningEvents\Lifecycle\Module;
+
+use Gacela\Framework\AbstractFactory;
+
+final class Factory extends AbstractFactory
+{
+    public function createGreeting(): string
+    {
+        return (string)$this->getProvidedDependency(Provider::GREETING);
+    }
+}